+++ CHANGELOG.rst
@@ -6,7 +6,7 @@ This document describes changes between each past release.
 2.2.0 (unreleased)
 ------------------
 
-- Nothing changed yet.
+- Use generic config keys as a fallback for missing specific signer config keys. (#151)
 
 
 2.1.1 (2017-10-27)